Do not return a connection with pending requests to the pool (#210)
* Close a connection with pending requests

A redis connection cannot return to a pool if it has pending requests, especially commands like "BLPOP/BRPOP". The command may block for an indefinite time, and stops the following commands from executing if the connection is acquired from the pool. Also, because the command is still executing, the BRPOP command may pop an item and drop it which creates bugs which are very hard to debug.

* test closing connection with pending requests

* fix test

* bug fix: use loop in input

* bug fix: parameters

* format