Support for MaxClientsError for connection and pool #186
When a new connection tries to get connected to the redis server and it
has been configured with a maximum number of connections, the client
will get a `MaxClientsError` exception if this limit is reached out.

Also if a pool initialization with a `min_size` that can meet the
requirements - beucase of that limit u others issues - the pool creation
will return also a `MaxClientsError`.