Use a custom StreamrReader #271
The connection opened with the server uses a customized StreamReader
that get rid of the double buffering between the operating system and
the Redis parser. This also allows aioredis to get rid of a current bug
with abruptly terminated connections.

Current implementation does not allow to keep giving support with
pause/resume the transport if some watermark has been overflowed,
the current implementation of hiredis does not publish the length of the
buffer.