Remove loop arguments from connection classes
This makes aioredis more consistent with other libraries that have
removed the need to explicitly specify the loop to use. This has two
knock-on effects:

1. __del__ will now always use asyncio.get_event_loop(). But `__del__`
is broken anyway, for reasons I've explained in #930.

2. PubSub.run_in_thread is a synchronous function but needs to know the
event loop to run. I've updated it to take an explicit loop argument.
This makes it possible to construct the Redis object before the event
loop. I'm going to file a separate issue to remove this function anyway,
since aioredis should provide concurrency via asyncio tasks rather than
threads, and the one test function has been disabled because it is
broken.

Closes #1002.