Replace PubSub.run_in_thread with coroutine run
The port from redis-py brought along PubSub.run_in_thread, and it was
modified to run code on the event loop. However, threading is not the
appropriate concurrency model for aioredis, so I've extracted the logic
and moved it into a `run` coroutine.

Also remove the test that covered run_in_thread and replace it with a
couple of tests for `run`. They're not the strongest tests (just send
and process a single message), but will at least catch simple errors
like missing awaits.