Fix a number of type annotation issues
This is mainly trying to fix the type annotations on the public API, so
that users of aioredis can run mypy, rather than trying to make aioredis
pass mypy itself.

I've introduced a number of abstract types like KeyT, StreamIdT etc;
this both serves as documentation on what namespace particular arguments
below to, and allows the definitions to be changed if we need to alter
the types allowed for a particular semantic.

I've tried to be careful to use KeyT only for the top-level keyspace of
Redis, which means things like the keys within a hash table, set or zset
are not annotated with this. I've left them as `str` for now, which
should be easy to find; a future pass should replace them with something
more generic, possibly EncodableT since it makes sense to other types
like integers as hash table keys (while it can be done for the root
keyspace, it seems less likely since one would normally scope numeric
IDs in some way e.g. `user:123`).

Other changes to the type annotations:
- Replace all instances of Collection with Iterable, which is more
  generic (allows generator expressions)
- Where an argument is passed to list_or_args, annotate it as
  Union[X, Iterable[X]] rather than Collection[X], since list_or_args
  allows for a singleton to be passed instead of a list.
- For commands taking an absolute expiry time (expireat, pexpireat),
  require int rather than float.
- Allow the `ex` and `px` parameters to `set` to be timedeltas
- Fix a number of functions taking a list of keys that were annotated as
  taking a single `str` (e.g. sdiff, sinter)
- Allow stream IDs to be specified as int (one can pass just a
  millisecond timestamp).
- Tighten xclaim to require message_ids be a list or tuple rather than
  any Sequence, since that's what the documentation requires (and the
  implementation checks it).
- Allow xpending_range to take endpoints as strings (stream IDs have
  more general form than just integers).
- Fix zset commands that take score ranges to allow them to be floats or
  strings (scores are floats, and bounds for scores can have a `[` or
  `(` prefix to make them inclusive/exclusive).
- Fix `count` argument to `georadius` and `georadiusbymember` to be `int`, not `bool`
- Fix **kwargs to `subscribe` and `psubscribe` to be Callable, not
  EncodableT. Refining the type of the callable is left for another
  time.
- Restrict bitfield offsets to be int or str (in the case of
  `BitFieldOperations.incrby`, it previously only allowed str, the other
  operations allowed EncodableT).