Improve type annotations for Mapping arguments
Arguments specified as `Mapping[KeyT, EncodableT]` or similar were
problematic because Mapping is no covariant on the key type, which means
that passing a `Dict[KeyT, str]` would work fine, but a `Dict[str, str]`
would not because str != KeyT.

Instead of using a union, use a TypeVar for Mapping key types. This
causes the type checker to infer the actual key type, which must be one
of the types listed. So now `Dict[str, str]` will work, although, say,
`Dict[Union[str, bytes], str]` will not. That could potentially be
improved by defining a Protocol that is similar to Mapping (possibly a
subset that only allows iteration, not lookups), but there are
diminishing returns since I assume most callers will pass either str or
bytes but not a mix.

See #1008.